move constants to constants/constants
