+++ frontend/src/components/products/ProductDetailsCard.js
@@ -1,7 +1,7 @@
 import {
   DEFAULT_PRODUCT_IMAGE,
   ROUTES_SHOP_QUERY_CATEGORY_PREFIX,
-} from "#utils/constants"
+} from "#constants/constants"
 import { setDefaultImageByError } from "#utils/dataManipulation"
 import { titleCase } from "#utils/stringManipulation"
 import { Link } from "react-router-dom"
