+++ frontend/src/components/products/ProductsCrud/GetAllProducts.js
@@ -1,7 +1,7 @@
 import NotificationCard from "#components/ui/NotificationCard"
 import ToggleButton from "#components/ui/ToggleButton"
+import { SHORT_MESSAGE_TIMEOUT, STYLES } from "#constants/constants"
 import useLocaleContext from "#context/localeContext"
-import { SHORT_MESSAGE_TIMEOUT, STYLES } from "#utils/constants"
 import { getAllProductsList } from "#utils/products"
 import { useCallback, useEffect, useState } from "react"
 import EditableListProductData from "./EditableListProductData"
