+++ frontend/src/pages/Login.js
@@ -1,15 +1,15 @@
 import LanguageDropdown from "#components/ui/LanguageDropdown"
 import SubNavbar from "#components/ui/Navbar/SubNavbar"
 import NotificationCard from "#components/ui/NotificationCard"
-import useLocaleContext from "#context/localeContext"
-import useUserContext from "#context/userContext"
 import {
   API_USERS_CURRENT_PREFIX,
   LOGO_IMAGE,
   LONG_MESSAGE_TIMEOUT,
   REDIRECT_TIMEOUT,
   ROUTES_SIGN_UP,
-} from "#utils/constants"
+} from "#constants/constants"
+import useLocaleContext from "#context/localeContext"
+import useUserContext from "#context/userContext"
 import { scrollToTop } from "#utils/render"
 import { loginUser } from "#utils/users"
 import {
