+++ frontend/src/pages/OrderSummary.js
@@ -1,8 +1,8 @@
 import NotificationCard from "#components/ui/NotificationCard"
+import { DEFAULT_PRODUCT_IMAGE, SHIPPING_COST } from "#constants/constants"
 import useCartContext from "#context/cartContext"
 import useLocaleContext from "#context/localeContext"
 import useUserContext from "#context/userContext"
-import { DEFAULT_PRODUCT_IMAGE, SHIPPING_COST } from "#utils/constants"
 import { setDefaultImageByError } from "#utils/dataManipulation"
 import { camelCaseToTitleCase } from "#utils/stringManipulation"
 import { getSummaryData } from "#utils/users"
