+++ frontend/src/pages/Shipping.js
@@ -1,9 +1,9 @@
 import FormInput from "#components/ui/FormInput"
 import NotificationCard from "#components/ui/NotificationCard"
+import { ROUTES_CURRENT_ADMIN, STYLES } from "#constants/constants"
 import useCartContext from "#context/cartContext"
 import useLocaleContext from "#context/localeContext"
 import useUserContext from "#context/userContext"
-import { ROUTES_CURRENT_ADMIN, STYLES } from "#utils/constants"
 import { listenInputChangeAndSetDataObject } from "#utils/dataManipulation"
 import React, { useEffect, useState } from "react"
 import { Link, useNavigate } from "react-router-dom"
