+++ frontend/src/pages/SignUp.js
@@ -1,13 +1,13 @@
 import LanguageDropdown from "#components/ui/LanguageDropdown"
 import SubNavbar from "#components/ui/Navbar/SubNavbar"
 import CreateUserForm from "#components/users/UsersCrud/CreateUserForm"
-import useLocaleContext from "#context/localeContext"
 import {
   LOGO_IMAGE,
   ROUTES_LOGIN,
   ROUTES_SIGN_UP,
   ROUTES_SIGN_UP_ADMIN,
-} from "#utils/constants"
+} from "#constants/constants"
+import useLocaleContext from "#context/localeContext"
 import { scrollToTop } from "#utils/render"
 import {
   faChevronDown,
